recurrent edges

+++ CONTROL/IOstuff.cpp
@@ -19,11 +19,11 @@ void IOstuff::neatInfoToStream(std::ofstream& stream, NEAT& neat) {
         << "\n";
     stream << neat.numIn << " " << neat.numOut << "\n";
     stream << neat.fitness << "\n" << neat.gencopies.size() << "\n" << neat.recurrentGeneCopies.size() << "\n";
-    for (int i = 0; i < neat.gencopies.size(); i++) {
+    for (size_t i = 0; i < neat.gencopies.size(); i++) {
         stream << neat.gencopies[i].getFrom() << " " << neat.gencopies[i].getTo() << " " << neat.gencopies[i].enabled
             << " " << neat.gencopies[i].weight << " " << neat.gencopies[i].childNodes << "\n";
     }
-    for (int i = 0; i < neat.recurrentGeneCopies.size(); i++) {
+    for (size_t i = 0; i < neat.recurrentGeneCopies.size(); i++) {
         stream << neat.recurrentGeneCopies[i].getFrom() << " " << neat.recurrentGeneCopies[i].getTo() << " " << neat.recurrentGeneCopies[i].enabled
             << " " << neat.recurrentGeneCopies[i].weight << "\n";
     }
